Moving compatibility code.
